backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2.16
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==6.1.0
gunicorn==20.1.0
idna==3.6
isort==5.12.0
mccabe==0.7.0
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.4
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
